Remove TwoFactor::Configuration and switch to ENV-based 2FA and encryption key checks
