+++ spec/system/two_factor_login_spec.rb
@@ -1,8 +1,18 @@
-# frozen_string_literal: true
-
 require 'rails_helper'
+require 'climate_control'
 
 RSpec.describe 'Two-factor authentication login', type: :system do
+  around do |example|
+    ClimateControl.modify(
+      'ENABLE_2FA' => '1',
+      'ACTIVE_RECORD_ENCRYPTION_PRIMARY_KEY' => 'primary',
+      'ACTIVE_RECORD_ENCRYPTION_DETERMINISTIC_KEY' => 'det',
+      'ACTIVE_RECORD_ENCRYPTION_KEY_DERIVATION_SALT' => 'salt'
+    ) do
+      example.run
+    end
+  end
+
   let(:user) { FactoryBot.create(:user, :with_otp) }
 
   before do
